LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.2_White_Space/S7.2_A5_T5-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Expected an identifier but found '\u00A0' instead
S7.2_A5_T5

PASS Expected parsing failure

TEST COMPLETE
